implemented login signup functionality for navbar
