+++ src/App.vue
@@ -1,6 +1,7 @@
 <script setup>
 import NavBar from "./components/NavBar.vue";
 </script>
+
 <template>
   <header>
     <div class="wrapper">
