+++ src/views/HomeView.vue
@@ -1,15 +1,43 @@
 <script setup>
+import { ref } from 'vue';
+import LogInComp from "../components/LogInComp.vue";
+import SignUpComp from "../components/SignUpComp.vue";
+
+const selectedComponent = ref('B'); // Initial state to null
+
+function ShowLogIn() {
+  selectedComponent.value = 'A'; // Show ComponentA
+}
+
+function showSignUp() {
+  selectedComponent.value = 'B'; // Show ComponentB
+}
+
 </script>
 
 <template>
-	<main>
-		<h1>Home View!</h1>
+	<nav>
+		<button @click="showSignUp">Sign Up</button>
+		<button @click="ShowLogIn">Log In</button>
+	</nav>
+	<main class="sing-in-form">
+		<div v-if="selectedComponent === 'A'">
+			<LogInComp />
+		</div>
+		<div v-else-if="selectedComponent === 'B'">
+			<SignUpComp />
+		</div>
 	</main>
 </template>
 
 <style scoped>
-h1 {
-	font-size: 5rem;
-	color: #1D6BF5;
+
+.inputContainer {
+	width: 100%;
+	text-align: center;
+}
+
+.inputContainer input {
+	width: 100%;
 }
 </style>
